--story=1000085 --user=Lawliet 【修复】需求列表和需求计数接口的父需求ID类型错误

+++ Src/TAPD.CSharpSDK/HttpData/Stories/TAPDStoriesRequest.cs
@@ -169,7 +169,7 @@ namespace TAPD.CSharpSDK
         /// 父需求
         /// </summary>
         [TAPDPropertyName("parent_id")]
-        public Nullable<int> parentID { get; set; }
+        public Nullable<long> parentID { get; set; }
 
         /// <summary>
         /// 子需求
